// src/app/models/film.ts
export interface Film {
  'id': number;
  'title': string;
  'episode_id': string;
  'opening_crawl': string;
  'director': string;
  'producer': string;
  'release_date': string;
  'characters': string;
  'planets': [];
  'starships': [];
  'vehicles': [];
  'species': [];
  'created': string;
  'edited': string;
  'url': string;
}